+++ lib/ayadn/app.rb
@@ -194,11 +194,16 @@ module Ayadn
     option :index, aliases: "-i", type: :boolean, desc: Descriptions.options_index
     option :raw, aliases: "-x", type: :boolean, desc: Descriptions.options_raw
     option :extract, aliases: "-e", type: :boolean, desc: Descriptions.options_extract
+    option :messages, type: :boolean, desc: 'Search in messages.'
     option :users, type: :boolean, desc: 'Search for users.'
     option :channels, type: :boolean, desc: 'Search for channels.'
     option :annotations, type: :boolean, desc: 'Search for specific annotations in posts.'
     def search(*words)
-      Action.new.search(words.join(","), options)
+      if options[:messages]
+        Action.new.search(words, options)
+      else
+        Action.new.search(words.join(","), options)
+      end
     end
 
     desc "settings", "List current Ayadn settings (-sg)"