+++ lib/ayadn/endpoints.rb
@@ -181,6 +181,15 @@ module Ayadn
       "#{@posts_url}search?annotation_types=#{anno}&access_token=#{Settings.user_token}#{@options_list}"
     end
 
+    def search_messages channel_id, words, options
+      if options[:count]
+        @options_list = API.build_query(options)
+      else
+        @options_list = API.build_query({count: Settings.options[:counts][:search]})
+      end
+      "#{@channels_url}/messages/search?query=#{words}&channel_ids=#{channel_id}&access_token=#{Settings.user_token}#{@options_list}"
+    end
+
     def search_channels words, options
         @options_list = API.build_query({count: Settings.options[:counts][:search]})
       "#{@channels_url}search?q=#{words}&order=popularity&access_token=#{Settings.user_token}#{@options_list}"